CHAOS-3513: close the lint and typecheck gates the same way as `test`
Three coupled defects, in the order that matters.

1. A workflow that does not gate on itself. Neither lint.yml nor
   typecheck.yml listed `.github/workflows/**` in its `code` filter, so a
   PR touching only a workflow file set code=false, skipped the gated
   job, and the required check reported SUCCESS. This is not
   hypothetical: in merge_group runs 30597898813 / 30597912205 (lint)
   and 30597898822 / 30597912214 (typecheck), on 2026-07-31, the queue's
   diff held exactly one file -- .github/workflows/governance.yml -- and
   PR #1349, a change to the CI gates themselves, reached main with
   neither lint nor typecheck having run. The guard added for this is
   derived, not hand-listed: each workflow must be selected by its own
   filter, so a fourth gate inherits it for free.

2. No merge_group clause. Both gated jobs selected on
   `workflow_dispatch || code == 'true'`, while test.yml's test-matrix
   carries a merge_group clause precisely because dorny/paths-filter has
   no base/head diff to compute in the queue. Added to both. Unlike (1)
   I have not observed this one firing; it is belt-and-braces, and the
   two runs above are explained by (1) alone.

3. The aggregator rule itself, which is what converts either skip from a
   silent pass into a red gate. Both gates now come through the script
   CHAOS-3482 introduced, generalized from "test.yml's two jobs" to any
   gate's list of jobs: each job is declared as
   `<name>|<policy>|<result>`, where the policy names the `if:`
   expression the script mirrors. A test asserts each job's real `if:`
   against the policy its workflow claims, so the label cannot lie.

The script also now fails when handed no jobs at all, and when handed a
policy it does not model -- both are wiring mistakes that would
otherwise produce a green check for a measurement that never happened.

Evidence: the pre-fix lint rule, materialized from origin/main and
driven through the new table, gets 9 of 20 rows wrong (including the
merge-queue row matching the observed incident); the fixed script gets
0 of 20 for both gates, and still 0 of 32 for `test` -- the same rows
that pinned CHAOS-3482, unchanged, which is the differential check on
the generalization. 12 clause-level mutations of the script, each
`bash -n`-valid, are all killed by a named row. 10 workflow mutations
across the three files each fail their specific guard and only for the
gate mutated.
