+++ .github/workflows/typecheck.yml
@@ -30,6 +30,14 @@ jobs:
         with:
           filters: |
             code:
+              # CHAOS-3513: omitting this put an unlinted, untypechecked
+              # change into main. In merge_group runs 30597898813 and
+              # 30597912205 the only file in the queue's diff was
+              # .github/workflows/governance.yml, so the filter reported
+              # code=false, the gated job skipped, and the required check
+              # reported SUCCESS. The gates could be edited -- weakened --
+              # by a PR they then declined to run on.
+              - '.github/workflows/**'
               - 'src/**'
               - 'tests/**'
               - 'ci/**'
@@ -40,7 +48,12 @@ jobs:
 
   typecheck-mypy:
     needs: [changes]
+    # dorny/paths-filter has no PR base/head diff to compute in the merge
+    # queue and can report code=false there, which would skip this job and
+    # let the aggregator see a skip it cannot attribute. Always run in the
+    # queue, as test.yml's test-matrix already does.
     if: >-
+      github.event_name == 'merge_group' ||
       github.event_name == 'workflow_dispatch' ||
       needs.changes.outputs.code == 'true'
     runs-on: ubuntu-latest
@@ -69,21 +82,23 @@ jobs:
 
   typecheck:
     name: typecheck
+    # Must remain `always()`. Under `!cancelled()` (or any condition that lets
+    # this job itself skip) a cancelled run leaves the REQUIRED check *skipped*,
+    # and branch protection counts a skipped required check as satisfied --
+    # which moves the silent pass one level up instead of closing it.
     if: always()
-    needs: [typecheck-mypy]
+    needs: [changes, typecheck-mypy]
     runs-on: ubuntu-latest
     steps:
+      - uses: actions/checkout@3d3c42e5aac5ba805825da76410c181273ba90b1 # v7.0.1
+
       - name: Aggregate typecheck result
         env:
-          TYPECHECK_RESULT: ${{ needs.typecheck-mypy.result }}
+          GATE_NAME: typecheck
+          EVENT_NAME: ${{ github.event_name }}
+          CHANGES_RESULT: ${{ needs.changes.result }}
+          CHANGES_CODE: ${{ needs.changes.outputs.code }}
+          GATED_JOB_1: typecheck-mypy|path-filtered|${{ needs.typecheck-mypy.result }}
         run: |
-          echo "typecheck-mypy result: ${TYPECHECK_RESULT}"
-          case "${TYPECHECK_RESULT}" in
-            success|skipped)
-              echo "typecheck gate passed"
-              ;;
-            failure|cancelled|*)
-              echo "typecheck gate failed"
-              exit 1
-              ;;
-          esac
+          chmod +x ci/aggregate_gate_results.sh
+          ./ci/aggregate_gate_results.sh
